pyproject.toml
[project]
name = "signals"
version = "0.1.0"
description = "Newsletter digest pipeline — fetch, sanitize, and summarize with Claude"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
